+++ bot/cogs/help.py
@@ -1,16 +1,7 @@
 from typing import List, Tuple
 import nextcord
 from nextcord.ext import commands, menus
-
-
-class DelBtn(nextcord.ui.View):
-    def __init__(self):
-        super().__init__(timeout=None)
-
-    @nextcord.ui.button(label="Delete", style=nextcord.ButtonStyle.secondary, emoji="<:dustbin:949602736633167882>")  
-    async def stop(self, button: nextcord.ui.Button, interaction: nextcord.Interaction):
-        await interaction.message.delete()
-
+from utils.delbtn import DelBtn
 
 class HelpPageSource(menus.ListPageSource):
     """Page source for dividing the list of tuples into pages and displaying them in embeds"""
