+++ bot/utils/delbtn.py
@@ -1,0 +1,10 @@
+import nextcord
+
+# A new nextcord view
+class DelBtn(nextcord.ui.View):
+    def __init__(self):
+        super().__init__(timeout=None)
+
+    @nextcord.ui.button(label="Delete", style=nextcord.ButtonStyle.secondary, emoji="<:dustbin:949602736633167882>")  
+    async def stop(self, button: nextcord.ui.Button, interaction: nextcord.Interaction):
+        await interaction.message.delete()
